LPS-79489 Move to the proper package that matches with API
